+++ sydtest-hspec/src/Test/Syd/Hspec.hs
@@ -101,6 +101,7 @@ runImportedItem (ImportedItem Hspec.Item {..}) trs wrapper = do
                 )
   let testRunResultNumTests = Nothing
   let testRunResultNumShrinks = Nothing
+  let testRunResultRetries = Nothing
   let testRunResultGoldenCase = Nothing
   let testRunResultFailingInputs = []
   let testRunResultExtraInfo = Nothing